requirements.txt
langchain==0.3.25
langchain-core==0.3.65
langchain-community==0.3.24
langchain-openai==0.3.19
langchain-google-genai==2.1.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.128.0
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
grad-cam==1.5.5
h11==0.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-multipart==0.0.21
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
ttach==0.0.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
